LayoutTests/platform/wpe/imported/w3c/web-platform-tests/pointerevents/pointerlock/pointerevent_pointerlock_after_pointercapture-expected.txt

CONSOLE MESSAGE: TypeError: div1.requestPointerLock is not a function. (In 'div1.requestPointerLock()', 'div1.requestPointerLock' is undefined)
CONSOLE MESSAGE: TypeError: div1.requestPointerLock is not a function. (In 'div1.requestPointerLock()', 'div1.requestPointerLock' is undefined)
Pointer Events pointer lock test

Follow the test instructions with mouse. If you don't have the device skip it.

Test Description: This test checks that we release the exsiting pointer capture when any element in the page gets a pointer lock.
Press left button down on the green rectangle and hold it.
Move the mouse inside the green rectangle.
Click right button while keeping left button down
Keep moving the mouse inside the green rectangle.
Test passes if the pointer capture is released on the yellow rectangle when the green rectangle gets the pointer lock.


Harness Error (FAIL), message = TypeError: div1.requestPointerLock is not a function. (In 'div1.requestPointerLock()', 'div1.requestPointerLock' is undefined)

FAIL mouse no pointercapture while pointerlock assert_equals: document.pointerLockElement should be div1. expected (object) Element node <div id="div1" style="width:800px;height:250px;background... but got (undefined) undefined
